Split up config parser from Config.
