+++ src/main/java/com/athaydes/logfx/config/Config.java
@@ -5,11 +5,9 @@ import com.athaydes.logfx.concurrency.TaskRunner;
 import com.athaydes.logfx.data.LogLineColors;
 import com.athaydes.logfx.text.HighlightExpression;
 import com.athaydes.logfx.ui.Dialog;
-import com.athaydes.logfx.ui.FxUtils;
 import javafx.application.Platform;
 import javafx.beans.InvalidationListener;
 import javafx.beans.property.SimpleObjectProperty;
-import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.collections.ObservableSet;
 import javafx.geometry.Orientation;
@@ -24,44 +22,29 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Iterator;
 import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Set;
 import java.util.concurrent.CompletableFuture;
-import java.util.regex.PatternSyntaxException;
-import java.util.stream.Stream;
 
 import static java.util.stream.Collectors.joining;
-import static java.util.stream.Collectors.toList;
 
 public class Config {
 
     private static final Logger log = LoggerFactory.getLogger( Config.class );
 
     private final Path path;
-    private final SimpleObjectProperty<LogLineColors> standardLogColors;
-    private final ObservableList<HighlightExpression> observableExpressions;
-    private final ObservableSet<File> observableFiles;
-    private final SimpleObjectProperty<Orientation> panesOrientation;
-    private final ObservableList<Double> paneDividerPositions;
-    private final BindableValue<Font> font;
+    private final ConfigProperties properties;
 
     public Config( Path path, TaskRunner taskRunner ) {
         this.path = path;
-
-        font = new BindableValue<>( Font.font( FxUtils.isMac() ? "Monaco" : "Courier New" ) );
-        standardLogColors = new SimpleObjectProperty<>( new LogLineColors( Color.BLACK, Color.LIGHTGREY ) );
-        observableExpressions = FXCollections.observableArrayList();
-        observableFiles = FXCollections.observableSet( new LinkedHashSet<>( 4 ) );
-        panesOrientation = new SimpleObjectProperty<>( Orientation.HORIZONTAL );
-        paneDividerPositions = FXCollections.observableArrayList();
+        this.properties = new ConfigProperties();
 
         if ( path.toFile().exists() ) {
             readConfigFile( path );
         } else {
-            observableExpressions.add( new HighlightExpression( "WARN", Color.YELLOW, Color.RED ) );
+            properties.observableExpressions.add( new HighlightExpression( "WARN", Color.YELLOW, Color.RED ) );
         }
 
         // make this a singleton object so that it can be remembered when we try to run it many times below
@@ -72,196 +55,45 @@ public class Config {
         InvalidationListener listener = ( event ) ->
                 taskRunner.runWithMaxFrequency( updateConfigFile, 2000L );
 
-        standardLogColors.addListener( listener );
-        observableExpressions.addListener( listener );
-        observableFiles.addListener( listener );
-        panesOrientation.addListener( listener );
-        paneDividerPositions.addListener( listener );
-        font.addListener( listener );
+        properties.standardLogColors.addListener( listener );
+        properties.observableExpressions.addListener( listener );
+        properties.observableFiles.addListener( listener );
+        properties.panesOrientation.addListener( listener );
+        properties.paneDividerPositions.addListener( listener );
+        properties.font.addListener( listener );
     }
 
     public SimpleObjectProperty<LogLineColors> standardLogColorsProperty() {
-        return standardLogColors;
+        return properties.standardLogColors;
     }
 
     public ObservableList<HighlightExpression> getObservableExpressions() {
-        return observableExpressions;
+        return properties.observableExpressions;
     }
 
     public ObservableSet<File> getObservableFiles() {
-        return observableFiles;
+        return properties.observableFiles;
     }
 
     public SimpleObjectProperty<Orientation> panesOrientationProperty() {
-        return panesOrientation;
+        return properties.panesOrientation;
     }
 
     public ObservableList<Double> getPaneDividerPositions() {
-        return paneDividerPositions;
+        return properties.paneDividerPositions;
     }
 
     public BindableValue<Font> fontProperty() {
-        return font;
+        return properties.font;
     }
 
     private void readConfigFile( Path path ) {
         try {
             Iterator<String> lines = Files.lines( path ).iterator();
-            parseConfigFile( null, lines );
+            new ConfigParser( properties ).parseConfigFile( null, lines );
         } catch ( IOException e ) {
             Dialog.showConfirmDialog( "Could not read config file: " + path +
                     "\n\n" + e );
-        }
-    }
-
-    private void parseConfigFile( String currentLine, Iterator<String> lines ) {
-        String line;
-        if ( currentLine != null ) {
-            line = currentLine;
-        } else if ( lines.hasNext() ) {
-            line = lines.next();
-        } else {
-            return; // no lines left
-        }
-
-        switch ( line.trim() ) {
-            case "standard-log-colors:":
-                parseStandardLogColors( lines );
-            case "expressions:":
-                parseExpressions( lines );
-            case "files:":
-                parseFiles( lines );
-            case "gui:":
-                parseGuiSection( lines );
-        }
-    }
-
-    private void parseStandardLogColors( Iterator<String> lines ) {
-        while ( lines.hasNext() ) {
-            String line = lines.next();
-            if ( line.startsWith( " " ) ) {
-                String[] components = line.trim().split( " " );
-                if ( components.length == 2 ) {
-                    try {
-                        Color bkg = Color.valueOf( components[ 0 ] );
-                        Color fill = Color.valueOf( components[ 1 ] );
-                        standardLogColors.set( new LogLineColors( bkg, fill ) );
-                    } catch ( IllegalArgumentException e ) {
-                        logInvalidProperty( "standard-log-colors", "color", line, e.toString() );
-                    }
-                } else {
-                    logInvalidProperty( "standard-log-colors", "number of colors", line, null );
-                }
-            } else if ( !line.trim().isEmpty() ) {
-                parseConfigFile( line, lines );
-                break;
-            }
-        }
-    }
-
-    private void parseExpressions( Iterator<String> lines ) {
-        while ( lines.hasNext() ) {
-            String line = lines.next();
-            if ( line.startsWith( " " ) ) {
-                try {
-                    observableExpressions.add( parseHighlightExpression( line.trim() ) );
-                } catch ( IllegalArgumentException e ) {
-                    logInvalidProperty( "expressions", "highlight", line, e.toString() );
-                }
-            } else if ( !line.trim().isEmpty() ) {
-                parseConfigFile( line, lines );
-                break;
-            }
-        }
-    }
-
-    private void parseFiles( Iterator<String> lines ) {
-        while ( lines.hasNext() ) {
-            String line = lines.next();
-            if ( line.startsWith( " " ) ) {
-                observableFiles.add( new File( line.trim() ) );
-            } else if ( !line.trim().isEmpty() ) {
-                parseConfigFile( line, lines );
-                break;
-            }
-        }
-    }
-
-    private void parseGuiSection( Iterator<String> lines ) {
-        while ( lines.hasNext() ) {
-            String line = lines.next();
-            if ( line.startsWith( " " ) ) {
-                String[] parts = line.trim().split( "\\s+" );
-                if ( parts.length == 0 ) {
-                    logInvalidProperty( "gui", "?", "empty line", null );
-                } else switch ( parts[ 0 ] ) {
-                    case "orientation":
-                        if ( parts.length != 2 ) {
-                            logInvalidProperty( "gui", "orientation", line,
-                                    "Expected 2 parts, got " + parts.length );
-                        } else try {
-                            panesOrientation.set( Orientation.valueOf( parts[ 1 ] ) );
-                        } catch ( IllegalArgumentException e ) {
-                            logInvalidProperty( "gui", "orientation", parts[ 1 ],
-                                    "Invalid value for Orientation: " + e );
-                        }
-                        break;
-                    case "pane-dividers":
-                        if ( parts.length != 2 ) {
-                            logInvalidProperty( "gui", "pane-dividers", line,
-                                    "Expected 2 parts, got " + parts.length );
-                        } else try {
-                            String[] separators = parts[ 1 ].split( "," );
-                            paneDividerPositions.addAll( toDoubles( separators ) );
-                        } catch ( IllegalArgumentException e ) {
-                            logInvalidProperty( "gui", "pane-dividers", parts[ 1 ],
-                                    e.toString() );
-                        }
-                        break;
-                    case "font":
-                        if ( parts.length < 3 ) {
-                            logInvalidProperty( "gui", "font", line,
-                                    "Expected 3 or more parts, got " + parts.length );
-                        } else {
-                            double size = 12.0;
-                            try {
-                                size = Double.parseDouble( parts[ 1 ] );
-                            } catch ( NumberFormatException e ) {
-                                logInvalidProperty( "gui", "font", parts[ 1 ],
-                                        "Font size is invalid: " + parts[ 1 ] );
-                            }
-
-                            String[] fontNameParts = Arrays.copyOfRange( parts, 2, parts.length );
-                            String fontName = String.join( " ", fontNameParts );
-                            try {
-                                font.setValue( Font.font( fontName, size ) );
-                            } catch ( IllegalArgumentException e ) {
-                                logInvalidProperty( "gui", "font", parts[ 0 ],
-                                        "Invalid font name: " + fontName );
-                            }
-                        }
-                        break;
-                }
-            } else if ( !line.trim().isEmpty() ) {
-                parseConfigFile( line, lines );
-                break;
-            }
-        }
-    }
-
-    private static List<Double> toDoubles( String[] numbers ) {
-        return Stream.of( numbers )
-                .map( String::trim )
-                .map( Double::parseDouble )
-                .collect( toList() );
-    }
-
-    private static void logInvalidProperty( String section, String name,
-                                            String invalidValue, String error ) {
-        log.warn( "Invalid value for {} in section {}: '{}'",
-                name, section, invalidValue );
-        if ( error != null ) {
-            log.warn( "Error: {}", error );
         }
     }
 
@@ -272,22 +104,22 @@ public class Config {
      */
     private void dumpConfigToFile() {
         CompletableFuture<LogLineColors> standardLogColorsFuture = new CompletableFuture<>();
-        Platform.runLater( () -> standardLogColorsFuture.complete( standardLogColors.get() ) );
+        Platform.runLater( () -> standardLogColorsFuture.complete( properties.standardLogColors.get() ) );
 
         CompletableFuture<List<HighlightExpression>> expressionsFuture = new CompletableFuture<>();
-        Platform.runLater( () -> expressionsFuture.complete( new ArrayList<>( observableExpressions ) ) );
+        Platform.runLater( () -> expressionsFuture.complete( new ArrayList<>( properties.observableExpressions ) ) );
 
         CompletableFuture<Set<File>> filesFuture = new CompletableFuture<>();
-        Platform.runLater( () -> filesFuture.complete( new LinkedHashSet<>( observableFiles ) ) );
+        Platform.runLater( () -> filesFuture.complete( new LinkedHashSet<>( properties.observableFiles ) ) );
 
         CompletableFuture<Orientation> panesOrientationFuture = new CompletableFuture<>();
-        Platform.runLater( () -> panesOrientationFuture.complete( panesOrientation.get() ) );
+        Platform.runLater( () -> panesOrientationFuture.complete( properties.panesOrientation.get() ) );
 
         CompletableFuture<List<Double>> paneDividersFuture = new CompletableFuture<>();
-        Platform.runLater( () -> paneDividersFuture.complete( new ArrayList<>( paneDividerPositions ) ) );
+        Platform.runLater( () -> paneDividersFuture.complete( new ArrayList<>( properties.paneDividerPositions ) ) );
 
         CompletableFuture<Font> fontFuture = new CompletableFuture<>();
-        Platform.runLater( () -> fontFuture.complete( font.getValue() ) );
+        Platform.runLater( () -> fontFuture.complete( properties.font.getValue() ) );
 
         standardLogColorsFuture.thenAccept( logLineColors ->
                 expressionsFuture.thenAccept( expressions ->
@@ -347,69 +179,6 @@ public class Config {
             Dialog.showConfirmDialog( "Could not write config file: " + path +
                     "\n\n" + e );
         }
-    }
-
-    protected static HighlightExpression parseHighlightExpression( String line )
-            throws IllegalArgumentException {
-        if ( line.isEmpty() ) {
-            throw highlightParseError( "empty highlight expression", line );
-        }
-        if ( line.startsWith( " " ) || line.endsWith( " " ) ) {
-            throw highlightParseError( "highlight expression contains invalid spaces at start/end", line );
-        }
-
-        int firstSpaceIndex = line.indexOf( ' ' );
-        if ( firstSpaceIndex <= 0 ) {
-            throw highlightParseError( "fill color and regular expression not specified", line );
-        }
-        String bkgColorString = line.substring( 0, firstSpaceIndex );
-
-        Color bkgColor;
-        try {
-            bkgColor = Color.valueOf( bkgColorString );
-        } catch ( IllegalArgumentException e ) {
-            throw highlightParseError( "invalid background color: '" + bkgColorString + "'", line );
-        }
-
-        if ( line.length() > firstSpaceIndex + 1 ) {
-            line = line.substring( firstSpaceIndex + 1 ).trim();
-        } else {
-            throw highlightParseError( "fill color and regular expression not specified", line );
-        }
-
-        int secondSpaceIndex = line.indexOf( ' ' );
-        if ( secondSpaceIndex <= 0 ) {
-            throw highlightParseError( "regular expression not specified", line );
-        }
-
-        String fillColorString = line.substring( 0, secondSpaceIndex );
-        Color fillColor;
-        try {
-            fillColor = Color.valueOf( fillColorString );
-        } catch ( IllegalArgumentException e ) {
-            throw highlightParseError( "invalid fill color: '" + fillColorString + "'", line );
-        }
-
-        if ( line.length() > secondSpaceIndex + 1 ) {
-            line = line.substring( secondSpaceIndex + 1 ).trim();
-        } else {
-            line = "";
-        }
-
-        if ( line.isEmpty() ) {
-            throw highlightParseError( "regular expression not specified", line );
-        }
-
-        String expression = line;
-        try {
-            return new HighlightExpression( expression, bkgColor, fillColor );
-        } catch ( PatternSyntaxException e ) {
-            throw highlightParseError( "cannot parse regular expression '" + expression + "'", line );
-        }
-    }
-
-    private static IllegalArgumentException highlightParseError( String error, String line ) {
-        return new IllegalArgumentException( "Invalid highlight expression [" + error + "]: " + line );
     }
 
 }
